requirements.txt
# Web Framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
WTForms==3.0.1

# Data Analysis and Processing
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# Database
SQLAlchemy==2.0.21

# Development and Testing
pytest==7.4.2
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0

# Utilities
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7

# Data validation
jsonschema==4.19.1
cerberus==1.3.5

# Date and time handling
python-dateutil==2.8.2

# HTTP requests (for future API integrations)
requests==2.31.0

# Security
bcrypt==4.0.1
cryptography==41.0.4

# MySQL Database
mysql-connector-python==8.2.0